+++ relay-pii/src/attachments.rs
@@ -1,6 +1,7 @@
 use regex::bytes::RegexBuilder as BytesRegexBuilder;
 use regex::{Match, Regex};
 use relay_event_schema::processor::{FieldAttrs, Pii, ProcessingState, ValueType};
+use serde_json::Deserializer;
 use smallvec::SmallVec;
 use std::borrow::Cow;
 use std::iter::FusedIterator;
@@ -8,7 +9,7 @@ use utf16string::{LittleEndian, WStr};
 
 use crate::compiledconfig::RuleRef;
 use crate::regexes::{get_regex_for_rule_type, ReplaceBehavior};
-use crate::{utils, CompiledPiiConfig, Redaction};
+use crate::{utils, CompiledPiiConfig, JsonScrubVisitor, Redaction, ScrubViewHierarchyError};
 
 /// The minimum length a string needs to be in a binary blob.
 ///
@@ -342,8 +343,8 @@ struct WStrSegment<'a> {
 
 /// A PII processor for attachment files.
 pub struct PiiAttachmentsProcessor<'a> {
-    pub(crate) compiled_config: &'a CompiledPiiConfig,
-    pub(crate) root_state: ProcessingState<'static>,
+    compiled_config: &'a CompiledPiiConfig,
+    root_state: ProcessingState<'static>,
 }
 
 /// Which encodings to scrub for `scrub_bytes`.
@@ -511,6 +512,27 @@ impl<'a> PiiAttachmentsProcessor<'a> {
         } else {
             false
         }
+    }
+
+    /// Applies PII rules to the given JSON.
+    ///
+    /// This function will perform PII scrubbing using `serde_transcode`, which means that it
+    /// does not have to lead the entire document in memory but will rather perform in on a
+    /// per-item basis using a streaming approach.
+    ///
+    /// Returns a scrubbed copy of the JSON document.
+    pub fn scrub_json(&self, payload: &[u8]) -> Result<Vec<u8>, ScrubViewHierarchyError> {
+        let output = Vec::new();
+
+        let visitor = JsonScrubVisitor::new(self.compiled_config);
+
+        let mut deserializer_inner = Deserializer::from_slice(payload);
+        let deserializer = crate::transform::Deserializer::new(&mut deserializer_inner, visitor);
+
+        let mut serializer = serde_json::Serializer::new(output);
+        serde_transcode::transcode(deserializer, &mut serializer)
+            .map_err(|_| ScrubViewHierarchyError::TranscodeFailed)?;
+        Ok(serializer.into_inner())
     }
 }
